+++ routes/_app.tsx
@@ -1,4 +1,5 @@
 import type { PageProps } from "fresh";
+import Navbar from "@/components/Navbar.tsx";
 
 export default function App({ Component }: PageProps) {
   return (
@@ -10,7 +11,12 @@ export default function App({ Component }: PageProps) {
         <link rel="stylesheet" href="/styles.css" />
       </head>
       <body>
-        <Component />
+        <div class="max-w-4xl mx-auto mt-2-1">
+          <Navbar />
+          <div class="mt-2-1">
+            <Component />
+          </div>
+        </div>
       </body>
     </html>
   );